BE/requirements-freeze.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloud-sql-python-connector==0.9.3
cryptography==42.0.5
Flask==3.0.3
Flask-Cors==4.0.0
frozenlist==1.4.1
google==3.0.0
google-ai-generativelanguage==0.6.2
google-api-core==2.18.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud==0.34.0
google-generativeai==0.5.2
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.2
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
multidict==6.0.5
numpy==1.26.4
openapi==1.1.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
PyMySQL==1.1.0
pyparsing==3.1.2
PyYAML==6.0.1
referencing==0.35.0
requests==2.31.0
rpds-py==0.18.0
rsa==4.9
scipy==1.13.0
soupsieve==2.5
SQLAlchemy==2.0.29
tqdm==4.66.2
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
Werkzeug==3.0.2
yarl==1.9.4
